Finish renaming ssh-crypt to sshcrypt
I also realized that printf %2.3 is a mistake because the first
number is the total characters, not number of digits before the
decimal point.  I had intended to leave room for the version
number to go to double digits, so now that is possible since
the script name shrank, and without breaking ability to decode
older versions.